GVT-1470: Add CSP support to frontend
This is done using 2 webpack plugins. Most of the heavy labor is handled by CspHtmlWebpackPlugin, which auto-generates nonces for the JS bundle and CSS styles. In order for this to work on styles, however, it was required that CSS was split into a separate file from the bundle, because then the CSP plugin is able to calculate a nonce for it.