General: Fix bad distinctification in km post layer
In this case (unlike GVT-1948), there was no consequence to getting
the distinctification wrong, because km posts would have to hit a map
tile border exactly in order to be fetched through multiple map tiles
anyway.

The problem, as in the other case, is that Set does the usual sort of
comparison in JavaScript, which doesn't look inside objects.