GVT-1764: Implement frontend caching for projects
Previously the whole list of projects was loaded via an API request
every time the project selection dropdown was opened in the infra-model
edit page. Changes to the list of projects should be rare, so they can
be cached client (browser) side instead.

The projects cache will be invalidated immediately upon adding a new
project within the local browser and external updates are
checked every 15 seconds also for the projects cache.

Similarly it is unnecessary to fetch a single project's information
from the API if all project data is already cached in the browser.
The getProject function that is responsible for fetching a single
project's data was refactored to use the cached list of projects.

The previously used fetchProjects-function was renamed to getProjects
as this seems to be the more common naming scheme within the ui-codebase.
Usage of this function was modified accordingly.