GVT-2039: Return null instead of throwing when topology linked switch is not found
It was concluded that the getTopologySwitchJointDataHolder function shouldn't
throw when a switch is not found. Instead the switch topology joint validation
should be done during other validation steps.

Switches linked to a location track by topology are not available for the
getTopologySwitchJointDataHandler function when they have not been staged
for publication.