GVT-2036: Fix non-working unknown value assignment for specific infra model fields
Previously the UI allowed setting "unknown" values for
 * Plan phase
 * Decision phase
 * Measurement method (horizontal)

The unknown values were modeled as "undefined" within the infra model UI related code.
When an object with an "undefined" value for field is serialized to JSON, the field
is completely removed from the resulting serialized JSON string. This further meant
that the values were not passed to the backend, which used the elvis-operator to
set a default value for the fields that were not available from the API request.

Given that the values were already modeled as nullable in the backend codebase,
it was concluded that all values that are not passed to the API from the UI
should be treated as "unknown". These "unknown" values are stored as nulls
in the infra model database, so default values can simply be removed in order
to allow a user to set these field values as "unknown".