GVT-2390 & GVT-2392: Fix publication view for older publications
Previously the URL & Ratko status display only worked for the most
recent publications as the publication view component was trying to
find a specific publication id from the list of most recent publications
instead of querying the api by it.

Similar issue was fixed for the Ratko publication status. One of the
checks to display unpublished or failed status was using the most
recent publications list instead of data from the publication that
was displayed.