GVT-2624: Fix drawing km-numbers in vertical geometry diagram while still handling cursor position properly
The position cross/cursor indicator of the vertical geometry diagram was previously modified in GVT-2583
as it was inproperly drawn to the right side of the actual position of the mouse. While the fix made during
GVT-2583 did work for the position indicator, it also modified the positioning of the km-numbers drawn to
the bottom of the vertical geometry diagram resulting in them being drawn off or not drawn at all to the
vertical geometry diagram's svg.

The changes in this commit should...
* Keep the previous fix to the cursor position still in place
* Draw all other elements as before.

It should be noted that the container used for the svg has 8px padding (on all sides) which means
that the container is a little larger than the actual svg image being drawn which was probably
the cause of the original issue of the cursor indicator being drawn with an offset to the side.