GVT-2722 & GVT-2665: Refactor scheduled task running
This allows tasks to be enabled/disabled on a more granular level.
Previously it was for example not possible to only enable cache-
preloading for the integration-api container without also
enabling all other scheduled tasks, such as csv generation
which the integration-api cannot accomplish due to only having
read-only access to the database.